requirements.txt
# Core Data Science Libraries
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2
scipy==1.11.4

# Machine Learning & NLP
sentence-transformers==2.2.2
transformers==4.35.2
torch==2.1.1

# Vector Database
lancedb==0.3.4
pyarrow==13.0.0

# Text Processing
nltk==3.8.1

# Web Requests
requests==2.31.0

# Configuration Management
PyYAML==6.0.1

# Development & Data Exploration (Optional)
jupyter==1.0.0
matplotlib==3.8.2
seaborn==0.12.2

# Progress Bars
tqdm==4.66.1 